+++ src/django_leads/tasks/__init__.py
@@ -4,5 +4,6 @@
 
 from django_leads.tasks.import_tasks import enqueue_import, fail_stale_import_batches, import_csv, import_form_lead
 from django_leads.tasks.intel_tasks import analyse_intel
+from django_leads.tasks.retention_tasks import anonymise_inactive
 from django_leads.tasks.rotation_tasks import rotate_thread, rotate_unresponsive
 from django_leads.tasks.rule_tasks import evaluate_rules